WIP: feat(vcs): rename `github` references to `vcs`
* This commit includes only the renames as part of
https://github.com/inveniosoftware/invenio-github/issues/188. The
changes are mostly pure find-and-replace renames of `github` to `vcs`
without behaviour or functionality being affected. However, there are
some small exceptions as otherwise it would be hard to separate files
that contain a mix of name-only and functional changes.

* This commit on its own is UNRELEASABLE. We will merge multiple commits
related to the VCS upgrade into the `vcs-staging` branch and then merge
them all into `master` once we have a fully release-ready prototype. At
that point, we will create a squash commit.