WIP: feat(vcs): view handlers
* Updated the view handlers and the Jinja templates to accommodate the
new generic structure. The view handlers needed only relatively small
changes, but the Jinja templates had to have all references to 'GitHub',
the GitHub logo, and GitHub URLs replaced with provider-specific
references. The 'vocabulary' defined in the
`RepositoryServiceProviderFactory` specifies the terminology that should
be used for a given VCS, including the name/icon but also the word for
'repository' which is 'project' for GitLab.

* The `ext.py` file is also changed here to reflect the dynamic submenu
entries depending on the VCS providers registered in the config.

* This commit on its own is UNRELEASABLE. We will merge multiple commits
related to the VCS upgrade into the `vcs-staging` branch and then merge
them all into `master` once we have a fully release-ready prototype. At
that point, we will create a squash commit.